requirements.txt
Django==2.2.1
stripe==2.28.0
django-stdimage==4.0.1
Pillow==6.0.0
